20221023 baseline2

+++ nebula-rust/src/graph_client/session.rs
@@ -249,7 +249,7 @@ impl<'a> Session<'a> {
 
     #[inline]
     // INSERT EDGE e2 (name, age) VALUES "11"->"13"@1:("n1", 12);
-    pub async fn insert_edge_with_rank(&self, space_name: &str, edge_name: &str, kv: HashMap<String, String>, from_vertex: &str, to_vertex: &str, rank: i64){
+    pub async fn insert_edge_with_rank(&self, space_name: &str, edge_name: &str, kv: HashMap<String, String>, from_vertex: &str, to_vertex: &str, rank: i32){
         let mut query = String::from("use ");
         query += space_name;
         query += "; ";
